Refactoring text stream
Creating a controller that updates text stream, instead of text stream
grabbing updates from the model directly. This removes the necessity
for the classes: "UnpackerTextStream" and the like because now text
stream knows nothing of the underlying model giving out the updates.
That logic has been moved to the UnpackerDisplayController, which
gets updates from the unpacker, and then unpdates the model.