Added console progressbar
Added a progress bar controller that controls progress bar length,
and also changed unpacker to query package size directly instead of
saving the packages size when it was contructed. The reason I needed
to do this is because packages recalculate their size when excludes
are added. Therefore unpacker's size will not be update to date unless
it grabs the size directly.