Fixing bug with unpacker blocking GUI.
Unpacker was executing on the EDT. This is because the Install button
action is executed on the EDT, this notifies the UnpackerController which
begins unpacking. Because the install button is on the EDT, the rest of
those actions are also on the EDT. I've switched this up. Clicking a
button now creates a new thread to execute these actions so it is
done off the EDT thread. This also means repaint methods are working
to update GUI elements.