Changing where unpacking thread is made.
Instead of spawning a new thread to update all button controllers
UnpackerController spawns a new thread for the unpacker itself.
It's silly to spawn a new thread for button controllers because
they probably don't require a thread. Unpacking however is long
operation and requires a new thread so the GUI is not blocked.