Updating how PathInputPanel is built
Pass swing components into PIP instead of passing in the UIComponents
that wrap the swing components, and then having PIP get the swing
components in it's build method. Passing in the swing components
directly makes more sense because panels only need to know about
the swing components, not the UIComponent wrapper.