Moving next/prev buttons to installer class
The next and previous buttons should be a part of the installer view
class because they are a common panel element (an element that every
panel has). Therefore I've updated the way the Installer controls the
GUI Views. When it builds the a GUI panel it passes the panel's build
method a panel to put all of the elements into. It then displays this
as it's center panel, with the button panel below it. This allows the
Installer to have common GUI elements that all panels share.