Fix issues with blank pathInput text causing npe
Package unpack directory is only set once the user sets the text.
So if the user hits install without setting the text, then the unpack
directory is null. I've changed this so when a package is added to a
PathInputController, the controller initializes its value to be whatever
the default for the field is (currently blank).