Getting VaultPanel logic working
This commit involves adding splitting GUIPathcomponent into two classes
to allow for File and Directory variants. This is so VaultPanel has a
way to validate Keystore and vault paths (which the user specifies as
new files and not existing ones). The VaultController class which
listens in on the VaultPanel components is responsible for passing
information to the model.